# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/post/create/Post":{"post":{"operationId":"PostController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":""}}}},"/post/findAllPosts":{"get":{"operationId":"PostController_findAll","parameters":[],"responses":{"200":{"description":"200. Success. Returns properties ","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PostsDto"}}}}}}}}},"/post/find/{id}":{"get":{"operationId":"PostController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/post/update/{id}":{"patch":{"operationId":"PostController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}}}}},"/post/delete/{id}":{"delete":{"operationId":"PostController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/company":{"post":{"operationId":"CompanyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"CompanyController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/company/{id}":{"get":{"operationId":"CompanyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"CompanyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"CompanyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"tekab-pfa-api","description":"this description","version":"1.0","contact":{}},"tags":[{"name":"cats","description":""}],"servers":[],"components":{"schemas":{"CreatePostDto":{"type":"object","properties":{"id":{"type":"number","description":"id of the post"},"title":{"type":"string","description":"title of post"},"content":{"type":"string","description":"content of post"},"published":{"type":"boolean","description":"published of post"},"authorId":{"type":"number","description":"authorId of the post"}},"required":["id","title","content","published","authorId"]},"UpdatePostDto":{"type":"object","properties":{"id":{"type":"number","description":"id of the post"},"title":{"type":"string","description":"title of post"},"content":{"type":"string","description":"content of post"},"published":{"type":"boolean","description":"published of post"},"authorId":{"type":"number","description":"authorId of the post"}},"required":["id","title","content","published","authorId"]},"CreateCompanyDto":{"type":"object","properties":{}},"UpdateCompanyDto":{"type":"object","properties":{}}}}}